Resolved PR comments ,optimized the code repository
