--- src/bonus-task.js
@@ -1,14 +1,0 @@
-// import { averageMaxTemperature,averageLowestTemperature } from "./utilities/monthly-average.js";
-
-// export function ChartReportGenerator(weatherData) {
-//     let highestAverageTemp = averageMaxTemperature(weatherData);
-//     let lowestAverageTemp = averageLowestTemperature(weatherData);
-//     console.log(
-//         "\x1b[31m" +
-//         "*".repeat(highestAverageTemp) +
-//         " " +
-//         `${highestAverageTemp} C` +
-//         "\x1b[0m"
-//     );
-  
-// }
