+++ src/main.js
@@ -1,16 +1,15 @@
 import {
+    extremeValuesReport,
     monthlyAverageReportGenerator,
     printDailyTemperatureExtremes,
 } from "./reports/report-generator.js";
-import { getWeatherFilesByYear } from "./utilities/read-weatherdata-dir.js";
-import { parseYearData } from "./parsers/yearly-weather-parser.js";
-import { yearExtremeValues } from "./utilities/get-extreme-values.js";
-import { getWeatherData } from "./utilities/get-weather-data.js";
-import yargsParser from "yargs-parser";
+import { getWeatherFilesByYear,getWeatherData, } from "./data/weather-data-reader.js";
+import { parseYearData} from "./parsers/parser.js";
+import Yargs from "yargs-parser";
 
 export async function main() {
     // eslint-disable-next-line no-undef
-    const args = yargsParser(process.argv.slice(2));
+    const args = Yargs(process.argv.slice(2));
     let weatherData;
     let yearBasedFiles;
     let yearWeatherReadings;
@@ -23,13 +22,12 @@ export async function main() {
                 break
             case "c":
                 weatherData = await getWeatherData(args.c);
-                //chartReportGenerator(weatherData);
                 printDailyTemperatureExtremes(weatherData, args.c)
                 break
             case "e":
                 yearBasedFiles = await getWeatherFilesByYear(args.e);
                 yearWeatherReadings = await parseYearData(yearBasedFiles);
-                yearExtremeValues(yearWeatherReadings);
+                extremeValuesReport(yearWeatherReadings)
                 break
         }
     }
