+++ src/reports/report-generator.js
@@ -1,62 +1,78 @@
-import { maxTemperatures, minTemperatures } from "../utilities/chart-report-calculation.js";
 import {
-    averageLowestTemperature,
-    averageMaxTemperature,
-    averageMeanHumidity,
-} from "../utilities/monthly-average.js";
-import { getMonthName } from "../utilities/get-monthname.js";
+  maxTemperatures,
+  minTemperatures,
+  averageLowestTemperature,
+  averageMaxTemperature,
+  averageMeanHumidity,
+  yearExtremeValues,
+} from "../calculations/calculations.js";
+import { getMonthName, getDayOfMonth } from "../utilities/get-month-details.js";
 
 export function monthlyAverageReportGenerator(weatherData) {
-    console.log(`Highest Average : ${averageMaxTemperature(weatherData)} C `);
-    console.log(`Lowest Average : ${averageLowestTemperature(weatherData)} C`);
-    console.log(`Average Mean Humidity : ${averageMeanHumidity(weatherData)} %`);
+  console.log(`Highest Average : ${averageMaxTemperature(weatherData)} C `);
+  console.log(`Lowest Average : ${averageLowestTemperature(weatherData)} C`);
+  console.log(`Average Mean Humidity : ${averageMeanHumidity(weatherData)} %`);
 }
 
 export function chartReportGenerator(weatherData) {
-    let highestAverageTemp = averageMaxTemperature(weatherData);
-    let lowestAverageTemp = averageLowestTemperature(weatherData);
-    console.log(
-        "\x1b[31m" +
-        "*".repeat(highestAverageTemp) +
-        " " +
-        `${highestAverageTemp} C` +
-        "\x1b[0m"
-    );
-    console.log(
-        "\x1b[34m" +
-        "*".repeat(lowestAverageTemp) +
-        " " +
-        `${lowestAverageTemp} C` +
-        "\x1b[0m"
-    );
+  let highestAverageTemp = averageMaxTemperature(weatherData);
+  let lowestAverageTemp = averageLowestTemperature(weatherData);
+  console.log(
+    "\x1b[31m" +
+      "*".repeat(highestAverageTemp) +
+      " " +
+      `${highestAverageTemp} C` +
+      "\x1b[0m"
+  );
+  console.log(
+    "\x1b[34m" +
+      "*".repeat(lowestAverageTemp) +
+      " " +
+      `${lowestAverageTemp} C` +
+      "\x1b[0m"
+  );
 }
 
 export function printDailyTemperatureExtremes(weatherData, monthAndYear) {
-    let maxTemps = maxTemperatures(weatherData)
-    let minTemps = minTemperatures(weatherData)
-    let [year,month] = monthAndYear.split("/")
+  let maxTemps = maxTemperatures(weatherData);
+  let minTemps = minTemperatures(weatherData);
+  let [year, month] = monthAndYear.split("/");
 
-    console.log(`${getMonthName(Number(month))} ${year}`)
+  console.log(`${getMonthName(month)} ${year}`);
 
-    for (let value of maxTemps) {
-        if (value === "") continue
-        console.log(
-            "\x1b[31m" +
-            "*".repeat(value) +
-            " " +
-            `${value} C` +
-            "\x1b[0m"
-        );
-    }
+  for (let value of maxTemps) {
+    if (value === "") continue;
+    console.log(
+      "\x1b[31m" + "*".repeat(value) + " " + `${value} C` + "\x1b[0m"
+    );
+  }
 
-    for (let value of minTemps) {
-        if (value === "") continue
-        console.log(
-            "\x1b[34m" +
-            "*".repeat(value) +
-            " " +
-            `${value} C` +
-            "\x1b[0m"
-        );
-    }
+  for (let value of minTemps) {
+    if (value === "") continue;
+    console.log(
+      "\x1b[34m" + "*".repeat(value) + " " + `${value} C` + "\x1b[0m"
+    );
+  }
+}
+
+export function extremeValuesReport(yearWeatherReadings) {
+  let [maxHumidity, maxTemp, minTemp] = yearExtremeValues(yearWeatherReadings);
+
+  console.log(
+    ` Highest: ${maxTemp["maxtemp"]}C on ${getMonthName(
+      maxTemp["date"]
+    )} ${getDayOfMonth(maxTemp["date"])}`
+  );
+
+  console.log(
+    ` Lowest: ${minTemp["minTemp"]}C on ${getMonthName(
+      minTemp["date"]
+    )} ${getDayOfMonth(minTemp["date"])}`
+  );
+
+  console.log(
+    ` Humidity: ${maxHumidity["maxHumidity"]}% on ${getMonthName(
+      maxTemp["date"]
+    )} ${getDayOfMonth(maxHumidity["date"])} `
+  );
 }
