+++ src/utilities/get-month-details.js
@@ -1,0 +1,12 @@
+/* eslint-disable no-unused-vars */
+export function getMonthName(value) {
+  const date = new Date(value);
+  const monthName = date.toLocaleString("default", { month: "short" });
+  return monthName;
+}
+
+export function getDayOfMonth(value){
+  const date = new Date(value);
+   const dayOfMonth = date.getDate();
+   return dayOfMonth
+}
